ilm: lockspace: Fix lockspace creation and deletion
This patch fixes lockspace creation and deletion:

For the lockspace creation, it assigns the lockspace pointer for up level
function, so the lockspace pointer can be stored into client data
structure and later can be deference by the client.

For both the lockspace creation and deletion, the patch adds the code to
send the result to local socket after handling the request.

Signed-off-by: Leo Yan <leo.yan@linaro.org>