ilm: python: Add ILM python module
This patch adds python module interface file lib_ilm.i, which is used to
export ILM lib APIs (in C code) to python module, so that we can use
python testing framework for testing.

To simplify the implementation, we don't need to manually write the
wrapper for python module, and directly use SWIG to build python module.

Signed-off-by: Leo Yan <leo.yan@linaro.org>