ilm: client: Add suspend/resume operation
During the client is handling any request, it needs to suspend the
client, so the main process can stop to receive any data from client and
the client's associated thread can receive data properly.

After the client has finished to handle the request, it needs to resume
the itself so that the main process can poll client FD and receive
further incoming messages.

Signed-off-by: Leo Yan <leo.yan@linaro.org>