ilm: pthread_backend: Refine lock break handling
If an IDM has hosts with countdown -1UUL, it's better to not directly
remove it from host list, especially if this IDM has other hosts have
not expired, in this case the host with countdown -1 still can be
granted and the lock cannot be broken.

For this reason, this patch changes the code to check if there have any
hosts is not expiration and if this is the case then it will bail out
rather than remove the hosts with countdown -1.  Only after detecting
no any hosts without expiration, the hosts with countdown -1 will be
removed from host list and move forward to break the lock.

Signed-off-by: Leo Yan <leo.yan@linaro.org>