ilm: lib: Add function for changing host ID
This patch is to add function for changing host ID, this is useful for
debugging and testing.  When a host create connect with daemon, the host
ID is generated with UUID + PID, so even the host create multiple socket
connections, the UUID + PID is same for all connections.

For this reason, add a new API to allow manually set the host ID, this
is helpful if we create two socket connections, and use the API to set
different host ID for these two connections thus can emulate the two
different hosts and do testing for multiple hosts testing cases.

Signed-off-by: Leo Yan <leo.yan@linaro.org>