ilm: lock: Get host count and lock mode
Thought the pthread backend emulation and raid lock have implemented
host count and lock mode, but it misses to provide the interface to
command request.  And in the code, it has some incorrect logic, e.g. the
function only supported to get state only when the host has acquired the
IDM, the host without acquiring the IDM cannot read out the host count
and lock mode, this is not right.

This patch adds the interface for command request framework, and changed
the raid lock and idm pthread emulation to allow to read state from any
specified drive, even the host has not acquired the IDM.

Signed-off-by: Leo Yan <leo.yan@linaro.org>