ilm: lockspace: Drop unneeded log
When the client disconnects with IDM lock manager, it closes the socket
and thus function ilm_lockspace_terminate() is invoked.  For a normal
exiting, ilm_lockspace_terminate() doesn't need to do anything and just
simply bail out after detects the lockspace has been deleted, for this
case, it should not output any error message, otherwise it will
introduce confusion, so drop this log.

Signed-off-by: Leo Yan <leo.yan@linaro.org>