ilm: lock: Free lock when terminate it
Even the low level function returns failure when unlock an IDM, still
free lock data structure when terminate it, this because the lock data
structure is only used to manage lock and after release it, it will not
be used anymore.

On the other hand, even IDM lock manager fails to release the IDM in
drives, we don't need to maintain for this case, since the IDM will be
timeout in drives and thus drive firmware can grant it to other hosts.

Signed-off-by: Leo Yan <leo.yan@linaro.org>