ilm: raid_lock: Refactor state machine
This patch is to refactor state machine, it introduces more states to
track an IDM for a drive:

  IDM_INIT: IDM is in unlock state
  IDM_BUSY: IDM is busy and acquired by other hosts
  IDM_DUPLICATE: IDM is acquired by the same host twice
  IDM_LOCK: IDM is in locked state
  IDM_TIMEOUT: IDM is timeout.

So this patch introduces several helper functions to maintain the state
machine and allow the state can be moved forward to next state.  Using
this way, we can unify the state machine code for all lock's operations.

Signed-off-by: Leo Yan <leo.yan@linaro.org>