ilm: lock: Support async operations
This patch is to support async operations by the raid thread, so
every time when send SCSI request, it will combine a request data
structure and add the request onto list.  In the middle of processing,
the raid thread moves request to "process" list, and after response it
will move the request to two different lists: "response" list is used to
track all requests and their result for normal operations, and another
"renew" list is used to track renewal requests.

Signed-off-by: Leo Yan <leo.yan@linaro.org>