ilm: raid_lock: Distinguish the two lists when return result
This patch fixes the issue that the renew thread and normal request
thread use the same list for return result.  So after applying this
patch, the renew thread will use the renew list and other normal request
will be queued onto response list; this will be more safe to use two
lists, every list is working for a thread.

Signed-off-by: Leo Yan <leo.yan@linaro.org>