ilm: lockspace: Avoid race condition between unlock and renew
Unlock and renew two operations might run in two separate threads, this
is possible to cause the state transition confusion, e.g. after unlock
the lock, the state can be locked state since this is caused by renew
thread to enlarge the majority.

This patch adds two functions to start and stop lock in a lockspace,
thus this can stop to renew the lock when unlock it.

Signed-off-by: Leo Yan <leo.yan@linaro.org>