ilm: raid_lock: Return -1ULL when detect breaking lock
When break a mutex, it's not way to figure out what's the latest LVB in
the drive, since the host which granted the lock previously has no
chance to update LVB into drive firmware, so after break this mutex and
get the ownership for it, we returns -1ULL to up layer to indicate the
LVB has not been updated properly, the up layer program (e.g. lvmlockd)
can use this occasion to update metadata.

Signed-off-by: Leo Yan <leo.yan@linaro.org>