ilm: lock: Don't bail out if fail to read UUID
It's possible to fail to read out partition's UUID, for this case,
the function ilm_read_blk_uuid() returns error and it sets all zero
into UUID buffer.

When the lock layer detects error for UUID is not found, don't bail out
and continue to generate drive list.  Though this cannot ensure all
hosts to use the same sequence to achieve majority, essentially hope
this is just minor case and we can use random sleeping for majority
locking which can avoid contention for multiple hosts acquring lock with
reversed sequence.

Signed-off-by: Leo Yan <leo.yan@linaro.org>