ilm: raid_lock: Refactor async operation polling
If the previous request has been handled and removed for process list,
if still set the previous request's FD into polling FD array, when
the poll() returns, the corresponding request cannot be found properly
for the previous request and causes the segment fault.

This patch refactors polling async operations, in every loop it
generates FD array based on the latest process list, and will check if
fail to find the corresponding request handle, which will be skipped to
next FD handling.

Signed-off-by: Leo Yan <leo.yan@linaro.org>