ilm: drive: Refactor device list generation
Now the lock manager will create a global device list, the global device
list will maintain the duplicate paths to the same drive.

Afterwards, if any locking operation pass device list, it needs to parse
the device list and find out which block devices are using the duplicated
paths, the lock manager will generate a single drive path for this case
so that can avoid to acquire the mutex twice on the same drive.

Signed-off-by: Leo Yan <leo.yan@linaro.org>