ilm: lock: Refactor drive number
It's possible to fail to find drive, e.g. after unmap device mapper for
specific device, for this case, we need to allow the program to continue
to run to try to achieve the locking majority.

To tolerate this case, this patch uses three variables 'good_drive_num',
'fail_drive_num' and 'total_drive_num' to track the drive states, the
'good_drive_num' is used to present the drives can be found SG path
correctly, 'fail_drive_num' records the failed drive number,
'total_drive_num' is the total drive number and the majority must
compare with 'total_drive_num'.

Signed-off-by: Leo Yan <leo.yan@linaro.org>