ilm: drive: Support cached device mapping
Sometimes if the device mapper is removed and we cannot find block
device and SCSI device node based on the device mapper name.  For this
case, we can firstly cache the device mapping and later can retrieve the
device mapping's block name and SCSI device name, this can help to
resolve the issue for the hardware device is broken in the middle of
LVM operations.

This patch might introduce a risk is that if a PV is unmapped and then
afterwards it's mapped with the same mapping name but with a different
hardware, the cached device mapping might contain stale info.  For this
case, later we can try to explore the information from dmeventd to get
updated info, but need to investigate for the relevant mechanism.

Signed-off-by: Leo Yan <leo.yan@linaro.org>