ilm: raid_lock: Fix the failure for reading LVB
If drive has no spare memory, it's possible to fail to acquire lock.
For this case, it hits the condition that the drive state is INIT and
tries to read LVB from drive, the current code doesn't permit for this.

This patch changes to allow inquire LVB, and find the LVB with matched
the host ID and lock ID.

Signed-off-by: Leo Yan <leo.yan@linaro.org>