ilm: idm: Fix for mutex destroy
So far it's found the maximum output data block is up to 128, so change
the data block number for reflect this.  And the drive firmware might
don't output data block with 'dead' flag, for this case, only destroy
mutex for unlock state, this can avoid to wrongly take the bottom data
block as valid data block but actually the data block which has not been
filled by drive so doesn't contain valid data.

Signed-off-by: Leo Yan <leo.yan@linaro.org>