ilm: idm: Fix reading the lock mode
When read the lock mode, if the mutex entry is in timeout state, should
don't bail out and needs to continue to find other mutex data block
until find a mutex entry with valid host.  Otherwise, it will return the
unlock state for the timeout host, and ignore the valid lock mode for
other hosts.

Signed-off-by: Leo Yan <leo.yan@linaro.org>