ilm: raid_lock: Handle the stale locking
If a previous user has used the locking and forget to release it, we
still need a method to acquire the lock again, especially if the initial
host has been timeout.  So the question is how the initial host can
acquire the mutex after its ownership has been timeout.

The designed the flow is firstly to unlock the mutex and acquire the
mutex again.  This flow can work well if the unlock operation passes the
'correct' lock mode.  When at the second time the initial host tries to
acquire the mutex with different lock mode, the unlock operation will
fail, so this patch will detect for this case, and tries to destroy the
mutex with reversed lock mode (say if fail to unlock with exclusive
mode and return -EINVAL, that means the mutex in firmware has the
opponent mode 'shareable', vice versa).

Signed-off-by: Leo Yan <leo.yan@linaro.org>