ilm: raid_lock: Support multi paths
Since the drives might contain multi paths, this patch adds the "path"
and "path_idx" in the request.  By default, the pointer "path" is
assigned to the first path of drive, if the I/O failure occurs, the
pointer "path" will be assigned to the second path, and so on.  So this
gives the raid locking more chance to access the same drive through
different paths.

This patch enhances the function idm_raid_multi_issue(), when detects
I/O failure, it send out the sequential requests by using the backup
paths.  The change keeps simple as possible to avoid over complex logic
for support multi paths.

Signed-off-by: Leo Yan <leo.yan@linaro.org>