ilm: drive: Support dynamically adding and removing drive path
When the drive is hot-plugged in or out, the block device will be
dynamically added or removed from system.  Use udev monitor to detect
the event for block device, and add and remove the drive path based on
the udev event.

Furthermore, for safe multi-threading, uses mutex to protect the drive
list.

Signed-off-by: Leo Yan <leo.yan@linaro.org>