ilm: drive: Add drive version number
Since the drive can be reset at any time, it can be reset by software
method or hardware method, when the IDM lock manager detects the drive
change, it will add or remove node from the drive list.

To tell the locking module if the drive list has been altered, this
patch introduced the drive version with monotonic increment, so when
every time the drive list is changed (either removing or adding drive
node), the version number will be increment.

Signed-off-by: Leo Yan <leo.yan@linaro.org>