ilm: lock: Update the drive paths for lock
When execute the locking operations, it's possible that the drive paths
have been altered which is caused by drive reset.

To use the latest drive paths, the locking flow will detect if the drive
version number can match with the cached local copy, if these two values
are not same, then release the old drive paths and update to new drive
paths based on the latest drive list.

Signed-off-by: Leo Yan <leo.yan@linaro.org>