ilm: raid_lock: Handle for drive firmware updating
It's possible that the firmware is updated on the fly, after the
firmware has been updated, all mutexes entries will be lost.  For this
case, the lock space thread will fail to renew the mutexes since the
drive firmware has reset all mutexes.

We can utilize the lockspace thread to renew the mutexes and thus it can
acquire the mutex after the drive firmware has been upgraded.  This
patch is to handle this case, if the drive firmware reports -ENOENT, the
lock manager will change the drive state to "IDM_INIT", thus it will
re-acquire the mutex.

Signed-off-by: Leo Yan <leo.yan@linaro.org>