ilm: build: Support version number
This patch is to support version number.

A new plain file "VERSION" is introduced to maintain the latest version
number, now its latest value is v0.9.0, and this number should be
consistent with the latest git tag, so at this point the git repo's
latest tag is v0.9.0.  The VERSION file can be always used even when the
released source code doesn't contain git history (e.g. the .git folder
has been totally removed).

When build the code with .git folder, we can create the version number
with more detailed info with the latest tag, commit ID and the dirty
state.  So it will dynamically generate a version number with "git
describe" command.

Afterwards, when create a new git tag, should update the version number
in VERSION file as well.

Signed-off-by: Leo Yan <leo.yan@linaro.org>