ilm: log: Record all logs into circular array
The current code only records logs with specific priority, any logs less
than the priority will not be saved into the cirtular array.

This patch changes to record all logs into the circular array and it
will only output the logs with checking the log's priority.  Saving all
logs into the circular array is a preparasion for later patch, which
will rollback the circual array if detect any error.

Signed-off-by: Leo Yan <leo.yan@linaro.org>