ilm: log: Support replaying logs
Now when set log level as debug level, it outputs verbose logs and
causes big file to save the logs.  To improve the log handling, it's
better to output warning and errors by default, when detect an error
occurs, it replays the logs so can give more info for debugging.

This patch is to support replaying logs when detect an error.

Signed-off-by: Leo Yan <leo.yan@linaro.org>