ilm: test: Change LBA for sg_raw command
In Seagate internal testing, it reports the error that LBA is out of
range.

The IDM specification says the LBA will be ignored in SCSI command,
but if pass big LBA value, it might cause confusion in Linux driver
or drive firmware.  To avoid error, using 0 for LBA in IDM SCSI
commands.

Signed-off-by: Leo Yan <leo.yan@linaro.org>