ilm: idm: Return success if without mutex entries
If the drive firmware returns success but there have no any mutex
entries, it should return success and fill the suitable values into
passed pointers.

For the asnyc operations, since we need to support the async result,
so change to data block number to 1 and defer to return the results
in the async result handler.

Signed-off-by: Leo Yan <leo.yan@linaro.org>